Update and rename mask_pywinlogs.py to tmask_pywinlogs.py

+++ tmask_pywinlogs.py
@@ -26,25 +26,25 @@ def checkEventLogSecurity(id):
     print(file)
     removeOldFile(str(old))
     if os.path.isfile(str(file)):
-        move(str(file),str(old))
+        move(str(file), str(old))
     d = []
     for log in c.Win32_NTLogEvent(EventCode=id, Logfile='Security'):
         data = re.sub('[^a-zA-Z0-9 \n\.\[\]\(\)\{\}]', '', str(log))
         with open(file, mode='a+', encoding='UTF-8', errors='strict', buffering=1) as plik:
             d = data.split('\n')
-            dd = (d[4],d[5],d[9],d[10],d[11],d[14],d[16])
+            dd = (d[4], d[5], d[9], d[10], d[11], d[14], d[16])
             plik.write(str(dd))
             plik.write('\n')
-            
-    with open(file, mode='r') as plik:  
-      with open(old, mode='r') as plik2:
-        difference = set(plik).difference(plik2)
-    difference.discard('\n')      
-    
-    with open(diff, mode='w') as plik_out:
-      for line in difference:
-            plik_out.write(line)
-    
+
+    if os.path.isfile(str(old)):
+        with open(file, mode='r') as plik:
+            with open(old, mode='r') as plik2:
+                difference = set(plik).difference(plik2)
+        difference.discard('\n')
+
+        with open(diff, mode='w') as plik_out:
+            for line in difference:
+                plik_out.write(line)
 
 
 def checkEventLogApplication():
@@ -55,26 +55,26 @@ def checkEventLogApplication():
     print(file)
     removeOldFile(str(old))
     if os.path.isfile(str(file)):
-        move(str(file),str(old))
+        move(str(file), str(old))
     d = []
     for log in c.Win32_NTLogEvent(Type="error", Logfile="Application"):
         data = re.sub('[^a-zA-Z0-9 \n\.\[\]\(\)\{\}]', '', str(log))
         with open(file, mode='a+', encoding='UTF-8', errors='strict', buffering=1) as plik:
             d = data.split('\n')
             #dd = d[0:]
-            dd = (d[4],d[5],d[9],d[10],d[14],d[16])
+            dd = (d[4], d[5], d[9], d[10], d[14], d[16])
             plik.write(str(dd))
             plik.write('\n')
-            
-            
-    with open(file, mode='r') as plik:  
-      with open(old, mode='r') as plik2:
-        difference = set(plik).difference(plik2)
-    difference.discard('\n')      
-    
-    with open(diff, mode='w') as plik_out:
-      for line in difference:
-            plik_out.write(line)
+
+    if os.path.isfile(str(old)):
+        with open(file, mode='r') as plik:
+            with open(old, mode='r') as plik2:
+                difference = set(plik).difference(plik2)
+        difference.discard('\n')
+
+        with open(diff, mode='w') as plik_out:
+            for line in difference:
+                plik_out.write(line)
 
 
 def checkEventLogSystem():
@@ -83,31 +83,31 @@ def checkEventLogSystem():
     old = (f'sys_err_logs_old.txt')
     diff = (f'sys_err_logs_diff.txt')
     print(file)
-    #removeOldFile(str(old))
-            
+    # removeOldFile(str(old))
+
     if os.path.isfile(str(file)):
-        move(str(file),str(old))
-        
+        move(str(file), str(old))
+
     d = []
-    
+
     for log in c.Win32_NTLogEvent(Type="error", Logfile="System"):
         data = re.sub('[^a-zA-Z0-9 \n\.\[\]\(\)\{\}]', '', str(log))
         with open(file, mode='a+', encoding='UTF-8', errors='strict', buffering=1) as plik:
             d = data.split('\n')
             dd = d[4:]
-            dd = (d[4],d[11],d[12],d[13])
+            dd = (d[4], d[11], d[12], d[13])
             plik.write(str(dd))
             plik.write('\n')
-            
-    with open(file, mode='r') as plik:  
-      with open(old, mode='r') as plik2:
-        difference = set(plik).difference(plik2)
-    difference.discard('\n')      
-    
-    with open(diff, mode='w') as plik_out:
-      for line in difference:
-            plik_out.write(line)
 
+    if os.path.isfile(str(old)):
+        with open(file, mode='r') as plik:
+            with open(old, mode='r') as plik2:
+                difference = set(plik).difference(plik2)
+        difference.discard('\n')
+
+        with open(diff, mode='w') as plik_out:
+            for line in difference:
+                plik_out.write(line)
 
 
 if __name__ == "__main__":
